Updated new strategy

+++ src/main/java/pk/Percentage.java
@@ -1,8 +1,9 @@
 package pk;
 
 public class Percentage {
-    public static double stdout(int play, int games){
-        double num = ((play)/games)*100;
+    // Calculates the stdout after the simulation is done
+    public static int stdout(int play, int games){
+        int num = (play/games)*100;
         return num;
     }
 }
